apiflow_langgraph/pyproject.toml
[project]
name = "apiflow-langgraph"
version = "0.1.0"
description = "LangGraph-powered self-healing execution engine for APIFlow"
authors = [{ name = "zayimhan" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
serve = "apiflow_langgraph.server:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apiflow_langgraph"]
